Use mlkem-native as AWS-LC's ML-KEM implementation
This imports mlkem-native (https://github.com/pq-code-package/mlkem-native)
as the ML-KEM implementation used in AWS-LC, replacing the reference
implementation.

mlkem-native is a high-performance, high-assurance C90 implementation
of ML-KEM developed under the Post-Quantum Cryptography Alliance (PQCA)
and the Linux Foundation. It is a fork of the reference implementation
that AWS-LC previously relied on, and remains close to it. mlkem-native
is the default ML-KEM implementation in
[libOQS](https://github.com/open-quantum-safe/liboqs).

**Import Mechanism**

The mlkem-native source code is unmodified and imported using the
importer script crypto/fipsmodule/ml_kem/importer.sh; the details
of the import are in META.yml.

Future updates to the source tree would ideally happen through
a re-import of a different version of mlkem-native, though a temporary
change-log is conceivable, similar to how the changes from the
reference implementation were documented so far.

**Import Scope**

mlkem-native has a C-only version as well as native 'backends' in
AVX2 and Neon for high performance. This commit only imports the
C-only version. Integration of native backends will be done separately.

mlkem-native offers its own FIPS-202 implementation, including
fast versions of batched FIPS-202. However, this commit does not
import those, but instead provides glue-code around AWS-LC's own
FIPS-202 implementation. The path to leveraging the FIPS-202 performance
improvements in mlkem-native would be to integrate them directly
into [crypto/fipsmodule/sha](crytpo/fipsmodule/sha).

**Formal Verification**

All C-code imported in this commit is formally verified using the
C Bounded Model Checker ([CBMC](https://github.com/diffblue/cbmc/))
to be free of various classes of undefined behaviour, including
out-of-bounds memory accesses and arithmetic overflow; the latter
is of particular interest for ML-KEM because of the use of lazy
modular reduction for improved performance.

The heart of the CBMC proofs are function contract and loop annotations to
the C-code. Function contracts are denoted `__contract__(...)` clauses
and occur at the time of declaration, while loop contracts are denoted
`__loop__` and follow the `for` statement.

The function contract and loop statements are kept in the source,
but removed by the preprocessor so long as the CBMC macro is undefined.
Keeping them simplifies the import, and care has been taken to make
them readable to the non-expert, and thereby serve as precise
documentation of assumptions and guarantees upheld by the code.

The CBMC proofs are automatic and don't require further proofs scripts;
yet, they come with their own build system and toolchain dependencies,
which this commit does not attempt to import. See
[proofs/cbmc](https://github.com/pq-code-package/mlkem-native/tree/main/proofs/cbmc)
in the mlkem-native repository. Mid-term, however, CI infrastructure
should be setup that allows to import and check the CBMC proofs
as part of the AWS-LC CI.

**Performance**

It is expected that the ML-KEM performance with this commit
is comparable to that of the reference implementation. This is
because the mlkem-native's fast backends are not yet imported,
the FIPS-202 code remains that of AWS-LC, and mlkem-native is
otherwise close to the reference implementation.

**Technical aspects**

At the core, mlkem-native is currently a 'single-level' implementation of
ML-KEM: A build of the main source tree provides an implementation of exactly
one of ML-KEM-512/768/1024, depending on the MLKEM_K parameter.
This property is inherited from the ML-KEM reference implementation, while
AWS-LC's fork of the reference implementation has changed this behaviour
and passes the security level as a runtime parameter.

To build all security levels, level-specific sources are built 3 times, once
per security level, and linked with a single build of the level-independent
code. The single-compilation-unit approach pursued by AWS-LC makes this process
fairly simple since one merely needs to include the single-compilation-unit file
provided by mlkem-native three times, and configure it so that the level-independent
code is included only once. The final include moreover `#undef`'ines all macros
defined by mlkem-native, reducing the risk of name clashes with other parts of
crypto/fipsmodule/bcm.c.

**Outstanding**

mlkem-native does not yet offer the following two FIPS-related features:
- Pairwise Consistency Test (PCT)
- Stack zeroization.
Those will be added to mlkem-native shortly and the import updated.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>